+++ src/main/java/com/modive/dashboard/controller/PostDriveDashboardController.java
@@ -1,18 +1,14 @@
 package com.modive.dashboard.controller;
 
 import com.modive.dashboard.dto.DriveDashboardResponse;
-import com.modive.dashboard.dto.DriveDetailDto;
+import com.modive.dashboard.dto.detail.DriveDetailDto;
 import com.modive.dashboard.dto.DriveListDto;
-import com.modive.dashboard.entity.Drive;
-import com.modive.dashboard.entity.DriveDashboard;
 import com.modive.dashboard.enums.ScoreType;
 import com.modive.dashboard.service.PostDriveDashboardService;
-import com.modive.dashboard.service.TotalDashboardService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.List;
 
 @RestController
