+++ src/main/java/com/modive/dashboard/dto/detail/AttentionDetailDto.java
@@ -1,4 +1,4 @@
-package com.modive.dashboard.dto;
+package com.modive.dashboard.dto.detail;
 
 public class AttentionDetailDto extends DriveDetailDto {
     public SubScoreDto drivingTime;
