+++ src/main/java/com/modive/dashboard/dto/detail/DriveDetailDto.java
@@ -1,4 +1,4 @@
-package com.modive.dashboard.dto;
+package com.modive.dashboard.dto.detail;
 
 public class DriveDetailDto {
     public double score;
