+++ src/main/java/com/modive/dashboard/dto/detail/EcoDetailDto.java
@@ -1,4 +1,4 @@
-package com.modive.dashboard.dto;
+package com.modive.dashboard.dto.detail;
 
 public class EcoDetailDto extends DriveDetailDto {
     public SubScoreDto idling;
