+++ src/main/java/com/modive/dashboard/dto/detail/PreventDetailDto.java
@@ -1,4 +1,4 @@
-package com.modive.dashboard.dto;
+package com.modive.dashboard.dto.detail;
 
 public class PreventDetailDto extends DriveDetailDto {
     public SubScoreDto reaction;
